Review by Ketan Talaulikar
Ketan Talaulikar has entered the following ballot position for
draft-ietf-raw-technologies-16: No Objection

When responding, please keep the subject line intact and reply to all
email addresses included in the To and CC lines. (Feel free to cut this
introductory paragraph, however.)

Please refer to https://www.ietf.org/about/groups/iesg/statements/handling-ballot-positions/
for more information about how to handle DISCUSS and COMMENT positions.

The document, along with other ballot positions, can be found here:
https://datatracker.ietf.org/doc/draft-ietf-raw-technologies/

----------------------------------------------------------------------
COMMENT:
----------------------------------------------------------------------

Thanks to the authors, contributors, and the WG for the work put into this very
informative document.

I have a few comments/suggestions that I would like to share:

1) There is scope for refining the document text (considering US English). I am
not going into an editorial review mode due to lack of time, but I could
recommend some tools to help - please reach out offline, if interested.

Let us trust the rFC editor on this 2) Abstract:

Suggest to remove this sentence since, IMHO, it does not belong: "Those
technologies were selected as part of the RAW WG formation and listed in the WG
charter."

It was added that for another review! We need a rough consensus.
 3) Section 1:

CURRENT
Those technologies were selected as part of the WG formation and listed in the
WG charter.

SUGGEST
The technologies studied in this document were identified in the charter during
the DetNet WG formation or as part of subsequent re-charter (when the WG picked
up the work on RAW).

NOt exactly. That was the RAW charter, not that of DetNet.What about:"The technologies studied in this document were identified in the charter during the RAW WG formation and inherited by DetNet (when the WG picked up the work on RAW)." ?4) Section 1:

CURRENT
The purpose for this is to support work at RAW to justify and enable DetNet
work on those and possibly more compatible technologies.

SUGGEST
The purpose of this document is to support and enable work on the these (and
possibly other similar compatible technologies) at the IETF specifically in the
DetNet working group working on RAW.

done!
Many thanks Ketan!--
Pascal