Review by Éric Vyncke Éric Vyncke has entered the following ballot position for draft-ietf-raw-technologies-16: No Objection
When responding, please keep the subject line intact and reply to all
email addresses included in the To and CC lines. (Feel free to cut this
introductory paragraph, however.)

Please refer to https://www.ietf.org/about/groups/iesg/statements/handling-ballot-positions/
for more information about how to handle DISCUSS and COMMENT positions.

The document, along with other ballot positions, can be found here:
https://datatracker.ietf.org/doc/draft-ietf-raw-technologies/

----------------------------------------------------------------------
COMMENT:
----------------------------------------------------------------------

Due to time pressure, I have only browsed through the document (thanks for it
BTW, it seems well written and exhaustive).

Thanks to Ron Bonica for his INT directorate review at:
https://datatracker.ietf.org/doc/review-ietf-raw-technologies-15-intdir-telechat-bonica-2025-03-28/

About section 4.1 s/IEEE Std 802.11-2012/IEEE Std 802.11-2024/ ? Also, as
802.11-2024 is out, should the following text be updated ? `IEEEE Std 802.11ax
[IEEE Std 802.11ax] (approved in 2021), throughput, latency, and reliability
enhancements in P802.11be [IEEE 802.11be] (approved in 2024).` Should
informational references be added in this section ?

I believe Dave wanted to give a sense of history here.
Now, I do not believe the reader will want to read successively the thousands of pages of older 802.11 specs or that they are hard to obtain if he really wanted to.

Thanks for the LS with IEEE/3GPP about this document and thanks to the
shepherd's write-up for listing these LS.

Indeed

Unsure whether `Those technologies were selected as part of the RAW WG
formation and listed in the WG charter.` should be in the abstract.

OK. Removing by 2 to 1 since it was added for one review and then 2 complained about it!

About section 5, I find the following sentence self-contradicting, i.e., a kind
nudge to rewrite it `It was adopted and widely deployed in the last 10 years by
the major competing standards, Wireless HART and ISA100.11a.`

What about
"
                     It was used as a base for the major industrial
   wireless process control standards, Wireless HART and ISA100.11a.
"

Section 7, please expand ICAO (expanded later in 7.1) and CNS.

done

Section 12, unsure whether this document should have normative references.

Dunno, I was explained at some point that the practice is to use normative for the refs that are more critical for the understanding of the informational doc.Was that Mirja?
I'd love to get a final word on this.

--
Pascal